backend/src/services/lmsService.ts
import Course from "../models/course";
import LMSAccount from "../models/lmsAccount";

class LMSService {
  async findCourseById(courseId: string) {
    return await Course.findOne({ course_id: courseId });
  }

  createUserAccount(
    email: string,
    firstName: string,
    lastName: string
  ): LMSAccount {
    return LMSAccount.create(email, firstName, lastName);
  }

  enrollUserInCourse(
    email: string,
    courseId: string
  ): { success: boolean; redirectUrl?: string } {
    // Simulate successful enrollment
    return {
      success: true,
      redirectUrl: `https://lms.example.com/courses/${courseId}`,
    };
  }
}

export default LMSService;
